All bugs are fixed.
Bugs were related to the bisect function. Since we want to select the index where the key value is greater than or equal to the start key of the node, I used bisect left -1. But I ignored the fact that this function returns the index of the value if the value is found in the list. So, subtracting 1 caused a problem when a the searched key value presented in the list. Therefore, I changed the wrote another function called binary_search_list to be able to find proper place for the key in the list for insertion/selection operations.