Bug Fix
In the initializer of the fiting tree object, a dummy element is put. This operation is done before setting buffer error and error.
I was using the constants from constant file before but in the last commit I changed some of them to use the variables in the object.
This change was the source to this error and it is now fixed.